Solve "Caesar encryption protocol" exercise
I think my solution is good. Only thing that's wrong with it is the
bound_char helper method.

Delegating functions to a shared module is not only cool but really helpful.
With only a few necessary functions of a protocol implemented you can still
get access to a lot of functions that compose of these necessary functions.

Moved macros exercises to there own folder.