Add timeout to Rumbl.InfoSys.compute
Default is 5 secs but a different one can be passed in opts.

Sends itself a message after 5 seconds as a signal of running out of
time instead of using a timeout. It might not look like it but it still
gets all results that compute in time even if it would timeout. This is
due to the fact that all messages in the inbox are checked before the
timeout comes into effect.